Removed h5 file-locking turn-off
It turns out the file locking prohibition was not an issue with
singularity but with the file system I was testing it on. File locking
is necessary later in the pipeline for sqlite usage, so turning it off
here doesn't do any good. Catching this error and telling the user that
the problem is with their filesystem is on my TODO list.